python_chat/server/requirements.txt
python-etcd
Flask
flask-jwt-extended
passlib
pika